backend/requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.0

# Deep Learning - PINNED VERSIONS for compatibility
tensorflow==2.15.0
numpy==1.24.3
pillow==10.1.0

# Data Processing - Compatible with numpy 1.24.3
pandas==2.0.3

# Production WSGI Server
gunicorn==21.2.0

# Additional dependencies for TensorFlow
h5py==3.10.0

